+++ CHANGELOG.md
@@ -28,6 +28,7 @@ been removed.
 - System Overview dashboard uniformly uses SI measurements for bytes
 - Errors on requeued DeltaFiles due to missing flows were not properly marked
 - Pod status probe no longer reports "undefined method `any?' for nil:NilClass" when scaling deployments
+- Monitor correctly parses GraphQL errors
 
 ### Tech-Debt/Refactor
 - Do not store 0-byte files in minio